fixed path to background logo, organized styling
