+++ client/src/pages/SignUpPage.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Helmet } from 'react-helmet-async';
-import '../styles/SignUpPage.css';
+import '../styles/pages/SignUpPage.css';
 
 interface SignUpFormData {
 	username: string;
