+++ client/src/styles/components/Community.css
