+++ client/src/styles/components/NavBar.css
@@ -64,7 +64,7 @@ body::before {
 	top: 50%;
 	left: 50%;
 	transform: translate(-50%, -50%);
-	background: url('../assets/logo.webp') no-repeat center center;
+	background: url('../../../src/assets/logo.webp') no-repeat center center;
 	background-size: contain;
 	opacity: 0.2;
 	width: 50%;
