+++ client/src/styles/pages/DocumentationPage.css
@@ -1,4 +1,3 @@
-/* DocumentationPage.css */
 .docTxt {
   display: flex;
   justify-content: center;
