+++ client/src/styles/pages/MainPage.css
